Big rename: CrdtMerge -> FwHeadless
Also renamed SendReceiveConfig section of appsettings.json to
FwHeadlessConfig so we get all the renames done at once.
