+++ .github/workflows/deploy-branch.yaml
@@ -41,10 +41,10 @@ jobs:
       version: ${{ needs.set-version.outputs.version }}
       label-latest: false
 
-  build-crdtmerge:
-    name: Build CrdtMerge
+  build-fw-headless:
+    name: Build FwHeadless
     needs: [ set-version ]
-    uses: ./.github/workflows/lexbox-crdtmerge.yaml
+    uses: ./.github/workflows/lexbox-fw-headless.yaml
     with:
       version: ${{ needs.set-version.outputs.version }}
       label-latest: false
@@ -52,7 +52,7 @@ jobs:
   deploy:
     name: Deploy Develop
     uses: ./.github/workflows/deploy.yaml
-    needs: [ build-api, build-ui, build-hgweb, build-crdtmerge, set-version ]
+    needs: [ build-api, build-ui, build-hgweb, build-fw-headless, set-version ]
     secrets: inherit
     with:
       version: ${{ needs.set-version.outputs.version }}
