+++ .github/workflows/lexbox-fw-headless.yaml
@@ -1,4 +1,4 @@
-name: Build CrdtMerge
+name: Build FwHeadless
 
 # https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#on
 on:
@@ -14,11 +14,11 @@ on:
         default: false
 
 env:
-  IMAGE_NAME: ghcr.io/sillsdev/lexbox-crdtmerge
+  IMAGE_NAME: ghcr.io/sillsdev/lexbox-fw-headless
 
 
 jobs:
-  publish-crdtmerge:
+  publish-fw-headless:
     timeout-minutes: 60
     runs-on: ubuntu-latest
 
@@ -51,10 +51,10 @@ jobs:
         with:
           dotnet-version: '9.x'
       - name: Dotnet build
-        run: dotnet build backend/CrdtMerge/CrdtMerge.csproj
-      # TODO: Write CrdtMerge unit tests, probably based on existing sync tests
+        run: dotnet build backend/FwHeadless/FwHeadless.csproj
+      # TODO: Write FwHeadless unit tests, probably based on existing sync tests
       # - name: Unit tests
-      #   run: dotnet test backend/CrdtMerge/CrdtMerge.csproj --logger:"xunit;LogFileName={assembly}.results.xml" --results-directory ./test-results --filter "Category!=Integration&Category!=FlakyIntegration" --blame-hang-timeout 10m
+      #   run: dotnet test backend/FwHeadless/FwHeadless.csproj --logger:"xunit;LogFileName={assembly}.results.xml" --results-directory ./test-results --filter "Category!=Integration&Category!=FlakyIntegration" --blame-hang-timeout 10m
       # - name: Publish unit test results
       #   uses: EnricoMi/publish-unit-test-result-action@8885e273a4343cd7b48eaa72428dea0c3067ea98 # v2.14.0
       #   if: always()
@@ -91,7 +91,7 @@ jobs:
       - uses: docker/build-push-action@2cdde995de11925a030ce8070c3d77a52ffcf1c0 # v5.3.0
         with:
           context: backend
-          file: backend/CrdtMerge/Dockerfile
+          file: backend/FwHeadless/Dockerfile
           build-args: |
             APP_VERSION=${{ inputs.version }}
           push: ${{ !env.ACT && github.repository == 'sillsdev/languageforge-lexbox' }}
