+++ .github/workflows/release-pipeline.yaml
@@ -43,10 +43,10 @@ jobs:
       version: ${{ needs.set-version.outputs.version }}
       label-latest: true
 
-  build-crdtmerge:
-    name: Build crdtmerge
+  build-fw-headless:
+    name: Build fw-headless
     needs: [ set-version ]
-    uses: ./.github/workflows/lexbox-crdtmerge.yaml
+    uses: ./.github/workflows/lexbox-fw-headless.yaml
     with:
       version: ${{ needs.set-version.outputs.version }}
       label-latest: true
@@ -54,7 +54,7 @@ jobs:
   deploy:
     name: Deploy Staging
     uses: ./.github/workflows/deploy.yaml
-    needs: [ build-api, build-ui, build-hgweb, build-crdtmerge, set-version ]
+    needs: [ build-api, build-ui, build-hgweb, build-fw-headless, set-version ]
     secrets: inherit
     with:
       version: ${{ needs.set-version.outputs.version }}
