+++ backend/FwHeadless/appsettings.Development.json
@@ -1,5 +1,5 @@
 {
-  "SendReceiveConfig": {
+  "FwHeadlessConfig": {
     "ProjectStorageRoot": "../../hgweb/repos",
     "LexboxUrl": "http://localhost/",
     "LexboxUsername": "admin",
