+++ backend/FwHeadless/appsettings.json
@@ -1,5 +1,5 @@
 {
-  "SendReceiveConfig": {
+  "FwHeadlessConfig": {
     "LexboxUsername": null
   },
   "Logging": {
