+++ deployment/base/pvc.yaml
@@ -40,7 +40,7 @@ spec:
 apiVersion: v1
 kind: PersistentVolumeClaim
 metadata:
-  name: crdtmerge
+  name: fw-headless
   namespace: languagedepot
   labels:
     app.kubernetes.io/part-of: languagedepot
