+++ deployment/base/secrets.yaml
@@ -89,7 +89,7 @@ stringData:
 apiVersion: v1
 kind: Secret
 metadata:
-  name: crdtmerge
+  name: fw-headless
   namespace: languagedepot
 stringData:
   CRDT_MERGE_SEND_RECEIVE_USERNAME: ''
