+++ deployment/local-dev/kustomization.yaml
@@ -39,7 +39,7 @@ patches:
     path: delete-oauth-certs.yaml
 
   - path: app-config.yaml
-  - path: crdtmerge-deployment.patch.yaml
+  - path: fw-headless-deployment.patch.yaml
   - path: lexbox-deployment.patch.yaml
   - path: ui-deployment.patch.yaml
   - path: hg-repos-pvc.patch.yaml
