+++ deployment/staging/kustomization.yaml
@@ -7,7 +7,7 @@ resources:
 components:
   - ../init-repos
 patches:
-  - path: crdtmerge-deployment.patch.yaml
+  - path: fw-headless-deployment.patch.yaml
   - path: lexbox-deployment.patch.yaml
   - path: app-config.yaml
   - path: hg-repos-volume.yaml
